Refactor to get rid of cycle
